timeout: Update to v2.0 with renamed functions and upstreaming
* timeout.el (timeout-debouce, timeout-debouced-func)
(timeout-throttle, timeout-throttled-func): For compatibility with
the version of this package to be upstreamed into Emacs, change
the names of the functions in this library to match the Elisp
naming conventions.  The changes are:

timeout-throttle  -> timeout-throttled-func
timeout-throttle  -> timeout-throttled-func
timeout-throttle! -> timeout-throttle
timeout-throttle! -> timeout-throttle

Update version of package to v2 because of breaking change.

* README.org: Mention breaking change.